requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.17.0 --index-url https://download.pytorch.org/whl/cpu
Pillow==10.2.0
